refactor: extract AppState from server module
