+++ spog/api/src/sbom/search.rs
@@ -1,5 +1,5 @@
+use crate::app_state::AppState;
 use crate::search;
-use crate::server::AppState;
 use actix_web::{web, HttpResponse};
 use actix_web_httpauth::extractors::bearer::BearerAuth;
 use spog_model::search::PackageSummary;
