+++ spog/api/src/sbom/vuln.rs
@@ -1,6 +1,6 @@
+use crate::app_state::{AppState, ResponseError};
 use crate::error::Error;
 use crate::guac::service::GuacService;
-use crate::server::{AppState, ResponseError};
 use crate::service::v11y::V11yService;
 use crate::utils::spdx::find_purls;
 use actix_web::cookie::time;
